feat: migrate index.js to TS
This is the first cut, cleanup to come.

Fixes #7
