+++ package.json
@@ -40,6 +40,7 @@
     "ast-pretty-print": "^2.0.1",
     "babel-plugin-tester": "^10.0.0",
     "kcd-scripts": "^7.2.0",
+    "type-fest": "^0.20.2",
     "typescript": "^4.1.3"
   },
   "eslintConfig": {
